# shill-guard/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: ShillGuard API
  description: |
    ShillGuard 水军智能检测系统接口文档。
    
    ## 认证说明
    登录成功后获取 `token`，后续所有需要认证的接口请在请求头添加：
    ```
    Authorization: Bearer <token>
    ```
    网关会自动解析 token 并注入 `X-User-Id`、`X-User-Role` 头，微服务内部通过这两个头识别用户身份。
    
    ## 角色权限
    | role 值 | 说明 |
    |---------|------|
    | 0 | 普通用户 |
    | 1 | 版主 |
    | 2 | 管理员 |
    
    ## 统一响应格式
    所有接口均返回统一格式：
    ```json
    {
      "code": 200,
      "message": "操作成功",
      "data": {}
    }
    ```
  version: 1.0.0
  contact:
    name: ShillGuard Team
servers:
  - url: http://localhost:8080
    description: 本地开发环境（API 网关）
  - url: http://localhost:3001/api
    description: 前端代理（开发调试）
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 登录后获取的 JWT Token
  schemas:
...
    Result:
      type: object
      properties:
        code:
          type: integer
          description: 业务状态码
          example: 200
        message:
          type: string
          description: 响应消息
          example: 操作成功
        data:
          description: 响应数据（类型随接口不同而变化）
    ResultVoid:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              type: object
              nullable: true
              example: null

    # 分页结构
    PageInfo:
      type: object
      properties:
        records:
          type: array
          description: 数据列表
        total:
          type: integer
          format: int64
          description: 总记录数
          example: 128
        size:
          type: integer
          description: 每页大小
          example: 10
        current:
          type: integer
          description: 当前页码
          example: 1
        pages:
          type: integer
          description: 总页数
          example: 13

    # 登录请求
...
    PostCreateDTO:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
          description: 帖子标题（必填）
          example: 分享一个好用的摄影技巧
        content:
          type: string
          description: 帖子正文内容（必填）
          example: 今天分享一个超实用的构图技巧，三分法则...
        coverUrl:
          type: string
          nullable: true
          description: 封面图URL（由文件服务上传后获取）
          example: "http://localhost:9000/shill-guard/images/cover.jpg"
        mediaUrl:
          type: string
          nullable: true
          description: 媒体资源URL（视频/图组）
          example: "http://localhost:9000/shill-guard/videos/video.mp4"
        postType:
          type: integer
          description: 帖子类型（1=纯文字 2=图文 3=视频）
          enum: [1, 2, 3]
          example: 2
        topicTag:
          type: string
          nullable: true
          description: 话题标签
          example: 摄影

    # 帖子实体
    ContentPost:
      type: object
      properties:
        postId:
          type: integer
          format: int64
          description: 帖子ID
          example: 2001
        userId:
          type: integer
          format: int64
          description: 发布者用户ID
          example: 1001
        username:
          type: string
          description: 发布者用户名
          example: user01
        nickname:
          type: string
          description: 发布者昵称
          example: 摄影爱好者
        avatarUrl:
          type: string
          nullable: true
          description: 发布者头像URL
          example: "http://localhost:9000/shill-guard/avatars/a.jpg"
        title:
          type: string
          description: 帖子标题
          example: 分享一个好用的摄影技巧
        content:
          type: string
          description: 帖子内容
          example: 今天分享一个超实用的构图技巧...
        coverUrl:
          type: string
          nullable: true
          description: 封面图URL
          example: "http://localhost:9000/shill-guard/images/cover.jpg"
        mediaUrl:
          type: string
          nullable: true
          description: 媒体URL
        postType:
          type: integer
          description: 帖子类型（1=纯文字 2=图文 3=视频）
          example: 2
        topicTag:
          type: string
          nullable: true
          description: 话题标签
          example: 摄影
        viewCount:
          type: integer
          description: 浏览数
          example: 1024
        likeCount:
          type: integer
          description: 点赞数
          example: 88
        commentCount:
          type: integer
          description: 评论数
          example: 23
        isLiked:
          type: boolean
          description: 当前用户是否已点赞（需登录）
          example: false
        status:
          type: integer
          description: 状态（0=正常 1=删除）
          example: 0
        createdTime:
          type: string
          format: date-time
          description: 发布时间
          example: "2024-06-15T14:30:00"

    # 发表评论请求
...
    ContentComment:
      type: object
      properties:
        commentId:
          type: integer
          format: int64
          description: 评论ID
          example: 3001
        userId:
          type: integer
          format: int64
          description: 评论者用户ID
          example: 1001
        username:
          type: string
          description: 评论者用户名
          example: user01
        nickname:
          type: string
          description: 评论者昵称
          example: 小红
        avatarUrl:
          type: string
          nullable: true
          description: 评论者头像
          example: "http://localhost:9000/shill-guard/avatars/a.jpg"
        postId:
          type: integer
          format: int64
          description: 所属帖子ID
          example: 2001
        parentCommentId:
          type: integer
          format: int64
          nullable: true
          description: 父评论ID（null=一级评论）
          example: null
        replyToUserId:
          type: integer
          format: int64
          nullable: true
          description: 被回复用户ID
          example: null
        content:
          type: string
          description: 评论内容
          example: 写得真好，学到了！
        likeCount:
          type: integer
          description: 点赞数
          example: 5
        status:
          type: integer
          description: 状态（0=正常 1=删除）
          example: 0
        createdTime:
          type: string
          format: date-time
          description: 评论时间
          example: "2024-06-15T15:00:00"

    # 禁言记录实体
...
    ErrorResult:
      type: object
      properties:
        code:
          type: integer
          description: 错误码
          example: 401
        message:
          type: string
          description: 错误描述
          example: 未登录或Token已过期
        data:
          type: object
          nullable: true
          example: null

# ===================== 接口定义 =====================
paths:
...
  /api/content/posts:
    get:
      tags: [帖子]
      summary: 获取帖子列表（分页）
      description: |
        获取帖子列表，支持按话题标签筛选，按发布时间倒序排列。
        无需登录即可访问。
      operationId: getPostList
      parameters:
        - name: pageNum
          in: query
          required: false
          description: 页码（从1开始）
          schema:
            type: integer
            default: 1
            example: 1
        - name: pageSize
          in: query
          required: false
          description: 每页数量（最大50）
          schema:
            type: integer
            default: 10
            example: 20
        - name: topicTag
          in: query
          required: false
          description: 话题标签筛选（精确匹配）
          schema:
            type: string
            example: 摄影
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PageInfo'
                          - type: object
                            properties:
                              records:
                                type: array
                                items:
                                  $ref: '#/components/schemas/ContentPost'
              example:
                code: 200
                message: 操作成功
                data:
                  records:
                    - postId: 2001
                      userId: 1001
                      nickname: "摄影爱好者"
                      title: "分享一个好用的摄影技巧"
                      content: "今天分享一个超实用的构图技巧..."
                      coverUrl: "http://localhost:9000/shill-guard/images/cover.jpg"
                      topicTag: "摄影"
                      viewCount: 1024
                      likeCount: 88
                      commentCount: 23
                      createdTime: "2024-06-15T14:30:00"
                  total: 128
                  size: 10
                  current: 1
                  pages: 13

    post:
      tags: [帖子]
      summary: 发布帖子
      description: |
        发布新帖子，需要登录。
        若包含图片或视频，请先调用文件上传接口获取 URL，再填入 `coverUrl` 或 `mediaUrl`。
      operationId: createPost
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreateDTO'
            examples:
              纯文字帖子:
                value:
                  title: "今天的心情随笔"
                  content: "生活总是充满惊喜..."
                  postType: 1
                  topicTag: "生活"
              图文帖子:
                value:
                  title: "分享一个好用的摄影技巧"
                  content: "今天分享一个超实用的构图技巧..."
                  coverUrl: "http://localhost:9000/shill-guard/images/cover.jpg"
                  postType: 2
                  topicTag: "摄影"
      responses:
        '200':
          description: 发布成功，返回帖子ID
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data:
                        type: integer
                        format: int64
                        description: 新帖子的 postId
              example:
                code: 200
                message: 操作成功
                data: 2001
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
        '422':
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              example:
                code: 422
                message: 标题不能为空
                data: null
  /api/content/posts/{postId}:
    get:
      tags: [帖子]
      summary: 获取帖子详情
      description: 获取指定帖子的完整信息，同时浏览数 +1。无需登录。
      operationId: getPostDetail
      parameters:
        - name: postId
          in: path
          required: true
          description: 帖子ID
          schema:
            type: integer
            format: int64
            example: 2001
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ContentPost'
              example:
                code: 200
                message: 操作成功
                data:
                  postId: 2001
                  userId: 1001
                  nickname: "摄影爱好者"
                  title: "分享一个好用的摄影技巧"
                  content: "今天分享一个超实用的构图技巧，三分法则是..."
                  coverUrl: "http://localhost:9000/shill-guard/images/cover.jpg"
                  topicTag: "摄影"
                  viewCount: 1025
                  likeCount: 88
                  commentCount: 23
                  isLiked: false
                  createdTime: "2024-06-15T14:30:00"
        '404':
          description: 帖子不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              example:
                code: 2001
                message: 帖子不存在
                data: null

    delete:
      tags: [帖子]
      summary: 删除帖子
      description: |
        删除指定帖子。
        - 帖子发布者可以删除自己的帖子。
        - 管理员（role >= 2）可以删除任意帖子。
      operationId: deletePost
      security:
        - BearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          description: 帖子ID
          schema:
            type: integer
            format: int64
            example: 2001
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultVoid'
              example:
                code: 200
                message: 操作成功
                data: null
        '403':
          description: 无权删除（非作者且非管理员）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              example:
                code: 403
                message: 无权限访问
                data: null
  /api/content/posts/{postId}/like:
    post:
      tags: [帖子]
      summary: 点赞帖子
      description: 对指定帖子点赞，每个用户每个帖子只能点赞一次。需要登录。
      operationId: likePost
      security:
        - BearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          description: 帖子ID
          schema:
            type: integer
            format: int64
            example: 2001
      responses:
        '200':
          description: 点赞成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultVoid'
              example:
                code: 200
                message: 操作成功
                data: null
        '400':
          description: 已经点赞过
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              example:
                code: 2003
                message: 已经点赞过了
                data: null

    delete:
      tags: [帖子]
      summary: 取消点赞帖子
      description: 取消对指定帖子的点赞。需要登录。
      operationId: unlikePost
      security:
        - BearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          description: 帖子ID
          schema:
            type: integer
            format: int64
            example: 2001
      responses:
        '200':
          description: 取消点赞成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultVoid'

  # ==================== 评论模块 ====================
  /api/content/comments/post/{postId}:
    get:
      tags: [评论]
      summary: 获取帖子的评论列表（分页）
      description: 获取指定帖子下的一级评论列表，按时间排序。无需登录。
      operationId: getCommentsByPost
      parameters:
        - name: postId
          in: path
          required: true
          description: 帖子ID
          schema:
            type: integer
            format: int64
            example: 2001
        - name: pageNum
          in: query
          required: false
          description: 页码
          schema:
            type: integer
            default: 1
            example: 1
        - name: pageSize
          in: query
          required: false
          description: 每页数量
          schema:
            type: integer
            default: 20
            example: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PageInfo'
                          - type: object
                            properties:
                              records:
                                type: array
                                items:
                                  $ref: '#/components/schemas/ContentComment'
              example:
                code: 200
                message: 操作成功
                data:
                  records:
                    - commentId: 3001
                      userId: 1002
                      nickname: "小明"
                      content: "写得真好！"
                      likeCount: 5
                      createdTime: "2024-06-15T15:00:00"
                  total: 23
                  size: 20
                  current: 1
...